requirements.txt
PyPDF2
uvicorn
fastapi
python-multipart
sentence-transformers
faiss-cpu
python-dotenv
boto3
botocore
jwt
PyJWT
requests
